requirements.txt
fastapi 
uvicorn 
langchain 
langchain-community 
langchain-openai
langgraph 
sentence-transformers 
pypdf
python-dotenv
faiss-cpu
langchain-huggingface